feature: create get products endpoint
